+++ source/renderer/app/components/wallet/mnemonic-input/constants.ts
@@ -1,0 +1,2 @@
+// has to be the same as in mnemonic-input/MnemonicInput.scss
+export const COLUMNS_COUNT = 3;